[CSM-258] Split wallet's `Full.hs` module apart
